backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
docker>=7.0.0
google-generativeai>=0.4.1
chromadb>=0.4.24
python-dotenv>=1.0.1
tiktoken>=0.6.0
openai
